requirements.txt
tensorflow==2.17.1
keras-facenet==0.3.2
mtcnn==0.1.1
pillow
numpy
scipy
opencv-python
tensorflow-hub==0.14.0
scikit-learn
matplotlib
imageio
imutils==0.5.4
requests
psutil
